minimal_requirements.txt
Flask==3.1.0
pandas==2.2.3
google-genai==1.2.0
openpyxl==3.1.5